Make build_one use a `version` instead of `checkout_path` plus `target_path`.
`build_one` was taking paths of checkout and destination and was
compiling them without knowing which version it was. As I'd like in a
near future also build translated documentations, I'd need to know
which version is being compiled. As it's possible, from the version
number to build the paths that was given, I think it's a good idea to
pass the version number instead of the various paths.

As I needed to allow the version to be given, and as I wanted to test
it in my home directory, I used argparse to hold the default values
for the various paths still allowing me to give paths to /tmp/ or my
home directory.

With no argument, or with only `-q`, the behavior is the same as
before, so salt scripts and crontabs need to be modified.

I "dropped" the highly-probably unused possibility to build the doc
given two paths, it was probably usefull to test the script, but as
the script also has a lot of hardcoded values, it was very hard to
test. It's now simplier to test, but I still have to add parameters to
skip purging fastly, chowning to another group than docs (that may not
exist while testing) or a flag to switch between git and
mercurial. But to keep the patch "simple" I still not implemented
them.